phantom-seal/requirements.txt
cryptography>=42.0
oqs-python>=0.10.0
web3>=6.0
python-dotenv>=1.0
